refactor(desktop): preserve repository size budgets
Problem:
- The first PR revision added lines to two components already governed by repolint file-size ratchets.

Root cause:
- Reusable animation completion and transcript entrance identity logic remained in the oversized component files.

Fix:
- Move fail-open exit animation handling into the shared motion module.
- Move model-backed transcript entrance setup into the entrance hook module.

Verification:
- go run ./tools/repolint
- pnpm exec tsx src/__tests__/native-motion.test.tsx
- pnpm exec tsx src/__tests__/approval-animation.test.tsx
- pnpm test:typecheck
- pnpm lint:hooks
- pnpm build
